# openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "Cliftin API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/workouts": {
      "get": {
        "operationId": "listWorkouts",
        "summary": "List completed workouts",
        "tags": [
          "workouts"
        ],
        "description": "Returns workouts in deterministic startedAt/id order. Date filters are inclusive calendar dates in the server TZ; returned timestamps are UTC.",
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false
          },
          {
            "schema": {
              "format": "date",
              "type": "string"
            },
            "example": "2026-07-01",
            "in": "query",
            "name": "from",
            "required": false
          },
          {
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "minimum": 1,
              "type": "integer"
            },
            "in": "query",
            "name": "programId",
            "required": false,
            "description": "Program primary key."
          },
          {
            "schema": {
              "minimum": 1,
              "type": "integer"
            },
            "in": "query",
            "name": "routineId",
            "required": false,
            "description": "Routine primary key."
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "startedAt"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "-startedAt"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "sort",
            "required": false,
            "description": "Single sort field; - means descending. Defaults to -startedAt."
          },
          {
            "schema": {
              "format": "date",
              "type": "string"
            },
            "example": "2026-07-31",
            "in": "query",
            "name": "to",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "properties": {
                          "durationSeconds": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "id": {
                            "minimum": 1,
                            "type": "integer",
                            "example": 4001
                          },
                          "program": {
                            "anyOf": [
                              {
                                "additionalProperties": false,
                                "type": "object",
                                "properties": {
                                  "id": {
                                    "minimum": 1,
                                    "type": "integer"
                                  },
                                  "name": {
                                    "anyOf": [
                                      {
                                        "type": "string"
                                      },
                                      {
                                        "type": "null"
                                      }
                                    ]
                                  }
                                },
                                "required": [
                                  "id",
                                  "name"
                                ]
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "routine": {
                            "anyOf": [
                              {
                                "additionalProperties": false,
                                "type": "object",
                                "properties": {
                                  "id": {
                                    "minimum": 1,
                                    "type": "integer"
                                  },
                                  "name": {
                                    "anyOf": [
                                      {
                                        "type": "string"
                                      },
                                      {
                                        "type": "null"
                                      }
                                    ]
                                  }
                                },
                                "required": [
                                  "id",
                                  "name"
                                ]
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "startedAt": {
                            "anyOf": [
                              {
                                "format": "date-time",
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          }
                        },
                        "required": [
                          "durationSeconds",
                          "id",
                          "program",
                          "routine",
                          "startedAt"
                        ]
                      }
                    },
                    "nextCursor": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "items"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          },
                          "path": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "message",
                          "path"
                        ]
                      }
                    },
                    "message": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 400,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/workouts/latest": {
      "get": {
        "operationId": "getLatestWorkout",
        "summary": "Get the latest completed workout",
        "tags": [
          "workouts"
        ],
        "description": "Returns the newest completed workout using the same detail shape as direct workout lookup.",
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "lb"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "kg"
                  ]
                }
              ]
            },
            "example": "lb",
            "in": "query",
            "name": "unit",
            "required": false,
            "description": "Weight display unit. Defaults to lb."
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "durationSeconds": {
                      "anyOf": [
                        {
                          "type": "number"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "id": {
                      "minimum": 1,
                      "type": "integer",
                      "example": 4001
                    },
                    "program": {
                      "anyOf": [
                        {
                          "additionalProperties": false,
                          "type": "object",
                          "properties": {
                            "id": {
                              "minimum": 1,
                              "type": "integer"
                            },
                            "name": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "null"
                                }
                              ]
                            }
                          },
                          "required": [
                            "id",
                            "name"
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "routine": {
                      "anyOf": [
                        {
                          "additionalProperties": false,
                          "type": "object",
                          "properties": {
                            "id": {
                              "minimum": 1,
                              "type": "integer"
                            },
                            "name": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "null"
                                }
                              ]
                            }
                          },
                          "required": [
                            "id",
                            "name"
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "startedAt": {
                      "anyOf": [
                        {
                          "format": "date-time",
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "exercises": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "properties": {
                          "exerciseId": {
                            "anyOf": [
                              {
                                "description": "Exercise-definition primary key.",
                                "minimum": 1,
                                "type": "integer"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "id": {
                            "description": "Performed exercise-result primary key.",
                            "minimum": 1,
                            "type": "integer"
                          },
                          "name": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "sets": {
                            "type": "array",
                            "items": {
                              "additionalProperties": false,
                              "type": "object",
                              "properties": {
                                "id": {
                                  "minimum": 1,
                                  "type": "integer"
                                },
                                "isWarmup": {
                                  "description": "Always present, including when false.",
                                  "type": "boolean"
                                },
                                "reps": {
                                  "anyOf": [
                                    {
                                      "type": "number"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ]
                                },
                                "rpe": {
                                  "anyOf": [
                                    {
                                      "type": "number"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ]
                                },
                                "timeSeconds": {
                                  "anyOf": [
                                    {
                                      "type": "number"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ]
                                },
                                "volume": {
                                  "additionalProperties": false,
                                  "type": "object",
                                  "properties": {
                                    "unit": {
                                      "description": "Weight display unit. Defaults to lb.",
                                      "anyOf": [
                                        {
                                          "type": "string",
                                          "enum": [
                                            "lb"
                                          ]
                                        },
                                        {
                                          "type": "string",
                                          "enum": [
                                            "kg"
                                          ]
                                        }
                                      ],
                                      "example": "lb"
                                    },
                                    "value": {
                                      "anyOf": [
                                        {
                                          "type": "number"
                                        },
                                        {
                                          "type": "null"
                                        }
                                      ]
                                    }
                                  },
                                  "required": [
                                    "unit",
                                    "value"
                                  ]
                                },
                                "weight": {
                                  "additionalProperties": false,
                                  "type": "object",
                                  "properties": {
                                    "unit": {
                                      "description": "Weight display unit. Defaults to lb.",
                                      "anyOf": [
                                        {
                                          "type": "string",
                                          "enum": [
                                            "lb"
                                          ]
                                        },
                                        {
                                          "type": "string",
                                          "enum": [
                                            "kg"
                                          ]
                                        }
                                      ],
                                      "example": "lb"
                                    },
                                    "value": {
                                      "anyOf": [
                                        {
                                          "type": "number"
                                        },
                                        {
                                          "type": "null"
                                        }
                                      ]
                                    }
                                  },
                                  "required": [
                                    "unit",
                                    "value"
                                  ]
                                }
                              },
                              "required": [
                                "id",
                                "isWarmup",
                                "reps",
                                "rpe",
                                "timeSeconds",
                                "volume",
                                "weight"
                              ]
                            }
                          }
                        },
                        "required": [
                          "exerciseId",
                          "id",
                          "name",
                          "sets"
                        ]
                      }
                    }
                  },
                  "required": [
                    "durationSeconds",
                    "id",
                    "program",
                    "routine",
                    "startedAt",
                    "exercises"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          },
                          "path": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "message",
                          "path"
                        ]
                      }
                    },
                    "message": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 400,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": []
}